frontend/my-website/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
openai==1.3.7
qdrant-client==1.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
